% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_tools.R
\name{plot_pics}
\alias{plot_pics}
\title{Plotting image data}
\usage{
plot_pics(dat, labels = NULL, nrow = NULL, ncol = NULL)
}
\arguments{
\item{dat}{matrix or data.frame of image data with images represented by pixel-intensity stored row-wise}

\item{labels}{Labels to put on the facets.}

\item{nrow}{Number of rows of the facet_grid}

\item{ncol}{Number of columns of the facet grid}
}
\description{
Several images are plotted via facet_grid. Does only make sense for a sample of
the data with up to ~ 30 images.
}
\details{
Pixel intensities are not gathered the same way for all data sets. You may need
to add combinations of 'coord_flip', 'scale_x_reverse', 'scale_y_reverse'.
}
