% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_tools.R
\name{plotly_viz}
\alias{plotly_viz}
\title{3D visualization of embeddings}
\usage{
plotly_viz(x, ...)
}
\arguments{
\item{x}{Matrix with embedding coordinates or embedding object as return by 'embed'}

\item{...}{Further parameters to be passed to plot_ly}
}
\description{
S3 method to visualize embedding objects via plotly
}
