% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_shiny.R
\name{shiny_viz}
\alias{shiny_viz}
\title{Interactive visualization of embeddings using shiny}
\arguments{
\item{l_embs}{list of embedding objects as produced by 'embed'}

\item{funs}{matrix or data.frame containing the functional observations row-wise}

\item{grid}{vector of grid points, respectively arguments at which the functions have been evaluated}

\item{grouping}{vector indicating group membership. Used for color coding.}
}
\description{
Interactive visualization of embeddings using shiny
}
\details{
Visualizes up to four embeddings of a functional data set.
}
