% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedding.R
\name{embed}
\alias{embed}
\title{Embedding data}
\arguments{
\item{dist_mat}{Either an object of class dist or square matrix of pairwise distance of the data to be embedded}

\item{method}{Manifold method use to compute the embedding. One of "isomap" (isometric feature mapping), "umap"
(uniform manifold approximation and projection), "diffmap" (diffusion map), "mds" (multi-dimensional scaling), and
"tsne" (t-distributed stochastic neighborhood embedding.)}

\item{...}{Arguments to be passed to the embedding methods}
}
\value{
Returns an embedding object of class <method>.
}
\description{
Wrapper function for several embedding methods.
}
