% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_tools.R
\name{plotly_viz}
\alias{plotly_viz}
\title{S3 method to visualize embedding objects via plotly}
\usage{
plotly_viz(x, ...)
}
\arguments{
\item{...}{Further parameters to be passed to plot_ly}

\item{emb}{Matrix with embedding coordinates or embedding object as return by 'embed'}
}
\description{
S3 method to visualize embedding objects via plotly
}
