% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_shiny.R
\name{shiny_viz}
\alias{shiny_viz}
\title{Interactive visualization of embeddings}
\arguments{
\item{l_embs}{list of embedding objects as produced by 'embed'}

\item{funs}{matrix or data.frame containing the functional observations row-wise}

\item{grid}{vector of grid points, respectively arguments at which the functions have been evaluated}

\item{grouping}{vector indicating group membership. Used for color coding.}
}
\description{
Visualizes a functional data set and up to four embeddings of it using shiny.
}
\details{
The dotted black line indicates the global mean function. After selecting points
in the embedding(s) the corresponding functions and their mean(dashed black line)
will be plotted as well. Note, if you select several hundred points, plotting takes some time.
In addition, the indices of the selected functions are depicted.
}
